+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/FLTCamExposureTests.swift
@@ -17,9 +17,9 @@ final class FLTCamExposureTests: XCTestCase {
     let mockDeviceOrientationProvider = MockDeviceOrientationProvider()
 
     let configuration = CameraTestUtils.createTestCameraConfiguration()
-    configuration.captureDeviceFactory = { mockDevice }
+    configuration.captureDeviceFactory = { _ in mockDevice }
     configuration.deviceOrientationProvider = mockDeviceOrientationProvider
-    let camera = FLTCam(configuration: configuration, error: nil)
+    let camera = CameraTestUtils.createTestCamera(configuration)
 
     return (camera, mockDevice, mockDeviceOrientationProvider)
   }