+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/PhotoCaptureTests.swift
@@ -17,7 +17,7 @@ final class PhotoCaptureTests: XCTestCase {
   private func createCam(with captureSessionQueue: DispatchQueue) -> FLTCam {
     let configuration = CameraTestUtils.createTestCameraConfiguration()
     configuration.captureSessionQueue = captureSessionQueue
-    return FLTCam(configuration: configuration, error: nil)
+    return CameraTestUtils.createTestCamera(configuration)
   }
 
   func testCaptureToFile_mustReportErrorToResultIfSavePhotoDelegateCompletionsWithError() {
@@ -173,8 +173,8 @@ final class PhotoCaptureTests: XCTestCase {
     FLTDispatchQueueSetSpecific(captureSessionQueue, FLTCaptureSessionQueueSpecific)
     let configuration = CameraTestUtils.createTestCameraConfiguration()
     configuration.captureSessionQueue = captureSessionQueue
-    configuration.captureDeviceFactory = { captureDeviceMock }
-    let cam = FLTCam(configuration: configuration, error: nil)
+    configuration.captureDeviceFactory = { _ in captureDeviceMock }
+    let cam = CameraTestUtils.createTestCamera(configuration)
 
     let filePath = "test"
     let mockOutput = MockCapturePhotoOutput()