+++ packages/pigeon/lib/src/swift/templates.dart
@@ -37,12 +37,12 @@ protocol ${instanceManagerFinalizerDelegateName(options)}: AnyObject {
 String instanceManagerFinalizerTemplate(InternalSwiftOptions options) => '''
 // Attaches to an object to receive a callback when the object is deallocated.
 internal final class ${_instanceManagerFinalizerName(options)} {
-  private static let associatedObjectKey = malloc(1)!
+  internal static let associatedObjectKey = malloc(1)!
 
   private let identifier: Int64
   // Reference to the delegate is weak because the callback should be ignored if the
   // `InstanceManager` is deallocated.
-  private weak var delegate: ${instanceManagerFinalizerDelegateName(options)}?
+  internal weak var delegate: ${instanceManagerFinalizerDelegateName(options)}?
 
   private init(identifier: Int64, delegate: ${instanceManagerFinalizerDelegateName(options)}) {
     self.identifier = identifier
@@ -57,7 +57,11 @@ internal final class ${_instanceManagerFinalizerName(options)} {
   }
 
   static func detach(from instance: AnyObject) {
-    objc_setAssociatedObject(instance, associatedObjectKey, nil, .OBJC_ASSOCIATION_ASSIGN)
+    let finalizer = objc_getAssociatedObject(instance, associatedObjectKey) as? ${_instanceManagerFinalizerName(options)}
+    if let finalizer = finalizer {
+      finalizer.delegate = nil
+      objc_setAssociatedObject(instance, associatedObjectKey, nil, .OBJC_ASSOCIATION_ASSIGN)
+    }
   }
 
   deinit {
@@ -219,6 +223,10 @@ final class ${swiftInstanceManagerClassName(options)} {
   /// The manager will be empty after this call returns.
   func removeAllObjects() throws {
     lockQueue.sync {
+      let weakInstancesEnumerator = weakInstances.objectEnumerator()!
+      while let instance = weakInstancesEnumerator.nextObject() {
+        ${_instanceManagerFinalizerName(options)}.detach(from: instance as AnyObject)
+      }
       identifiers.removeAllObjects()
       weakInstances.removeAllObjects()
       strongInstances.removeAllObjects()