+++ script/tool/test/analyze_command_test.dart
@@ -16,12 +16,13 @@ void main() {
   late MockPlatform mockPlatform;
   late Directory packagesDir;
   late RecordingProcessRunner processRunner;
+  late RecordingProcessRunner gitProcessRunner;
   late CommandRunner<void> runner;
 
   setUp(() {
     mockPlatform = MockPlatform();
     final GitDir gitDir;
-    (:packagesDir, :processRunner, gitProcessRunner: _, :gitDir) =
+    (:packagesDir, :processRunner, :gitProcessRunner, :gitDir) =
         configureBaseCommandMocks(platform: mockPlatform);
     final AnalyzeCommand analyzeCommand = AnalyzeCommand(
       packagesDir,
@@ -469,5 +470,91 @@ void main() {
         ),
       ]),
     );
+  });
+
+  group('file filtering', () {
+    test('runs command for changes to Dart source', () async {
+      createFakePackage('package_a', packagesDir);
+
+      gitProcessRunner.mockProcessesForExecutable['git-diff'] =
+          <FakeProcessInfo>[
+        FakeProcessInfo(MockProcess(stdout: '''
+packages/package_a/foo.dart
+''')),
+      ];
+
+      final List<String> output =
+          await runCapturingPrint(runner, <String>['analyze']);
+
+      expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('Running for package_a'),
+          ]));
+    });
+
+    const List<String> files = <String>[
+      'foo.java',
+      'foo.kt',
+      'foo.m',
+      'foo.swift',
+      'foo.c',
+      'foo.cc',
+      'foo.cpp',
+      'foo.h',
+    ];
+    for (final String file in files) {
+      test('skips command for changes to non-Dart source $file', () async {
+        createFakePackage('package_a', packagesDir);
+
+        gitProcessRunner.mockProcessesForExecutable['git-diff'] =
+            <FakeProcessInfo>[
+          FakeProcessInfo(MockProcess(stdout: '''
+packages/package_a/$file
+''')),
+        ];
+
+        final List<String> output =
+            await runCapturingPrint(runner, <String>['analyze']);
+
+        expect(
+            output,
+            isNot(containsAllInOrder(<Matcher>[
+              contains('Running for package_a'),
+            ])));
+        expect(
+            output,
+            containsAllInOrder(<Matcher>[
+              contains('SKIPPING ALL PACKAGES'),
+            ]));
+      });
+    }
+
+    test('skips commands if all files should be ignored', () async {
+      createFakePackage('package_a', packagesDir);
+
+      gitProcessRunner.mockProcessesForExecutable['git-diff'] =
+          <FakeProcessInfo>[
+        FakeProcessInfo(MockProcess(stdout: '''
+README.md
+CODEOWNERS
+packages/package_a/CHANGELOG.md
+''')),
+      ];
+
+      final List<String> output =
+          await runCapturingPrint(runner, <String>['analyze']);
+
+      expect(
+          output,
+          isNot(containsAllInOrder(<Matcher>[
+            contains('Running for package_a'),
+          ])));
+      expect(
+          output,
+          containsAllInOrder(<Matcher>[
+            contains('SKIPPING ALL PACKAGES'),
+          ]));
+    });
   });
 }